test: SimilarToRecentlyViewedArtworks
